Move common data from local into lnk_info
