+++ LnkParse3/lnk_info.py
@@ -131,3 +131,113 @@ class LnkInfo:
         """
         start, end = 24, 28
         return unpack("<I", self._raw[start:end])[0]
+
+    def _has_opt_fields(self):
+        """
+        Offsets to the optional fields are specified.
+        """
+        return bool(self.header_size() >= 0x00000024)
+
+    def common_path_suffix(self):
+        """CommonPathSuffix (variable):
+        A NULL-terminated string, defined by the system default code page,
+        which is used to construct the full path to the link item or link
+        target by being appended to the string in the LocalBasePath field.
+        """
+        if not self.common_path_suffix_offset():
+            return None
+
+        start = self.common_path_suffix_offset()
+
+        binary = self._raw[start:]
+        text = self.text_processor.read_string(binary)
+        return text
+
+    def local_base_path_offset_unicode(self):
+        """LocalBasePathOffsetUnicode (4 bytes):
+        An optional, 32-bit, unsigned integer that specifies the location of
+        the LocalBasePathUnicode field. If the VolumeIDAndLocalBasePath flag
+        is set, this value is an offset, in bytes, from the start of the
+        LinkInfo structure; otherwise, this value MUST be zero. This field
+        can be present only if the value of the LinkInfoHeaderSize field is
+        greater than or equal to 0x00000024.
+        """
+        if not self._has_opt_fields():
+            return None
+
+        start, end = 28, 32
+        return unpack("<I", self._raw[start:end])[0]
+
+    def common_path_suffix_offset_unicode(self):
+        """CommonPathSuffixOffsetUnicode (4 bytes):
+        An optional, 32-bit, unsigned integer that specifies the location of
+        the CommonPathSuffixUnicode field. This value is an offset, in bytes,
+        from the start of the LinkInfo structure. This field can be present
+        only if the value of the LinkInfoHeaderSize field is greater than or
+        equal to 0x00000024.
+        """
+        if not self._has_opt_fields():
+            return None
+
+        start, end = 32, 36
+        return unpack("<I", self._raw[start:end])[0]
+
+    def local_base_path(self):
+        """LocalBasePath (variable):
+        An optional, NULL-terminated string, defined by the system default
+        code page, which is used to construct the full path to the link item
+        or link target by appending the string in the CommonPathSuffix field.
+        This field is present if the VolumeIDAndLocalBasePath flag is set.
+        """
+        if not self.local_base_path_offset():
+            return None
+
+        start = self.local_base_path_offset()
+
+        binary = self._raw[start:]
+        text = self.text_processor.read_string(binary)
+        return text
+
+    def common_network_relative_link(self):
+        """CommonNetworkRelativeLink (variable):
+        An optional CommonNetworkRelativeLink structure (section 2.3.2) that
+        specifies information about the network location where the link
+        target is stored.
+        """
+        # TODO:
+        pass
+
+    def local_base_path_unicode(self):
+        """LocalBasePathUnicode (variable):
+        An optional, NULL-terminated, Unicode string that is used to construct
+        the full path to the link item or link target by appending the string
+        in the CommonPathSuffixUnicode field. This field can be present only
+        if the VolumeIDAndLocalBasePath flag is set and the value of the
+        LinkInfoHeaderSize field is greater than or equal to 0x00000024.
+        """
+        if not self.local_base_path_offset_unicode():
+            return None
+
+        start = self.local_base_path_offset_unicode()
+
+        binary = self._raw[start:]
+        text = self.text_processor.read_unicode_string(binary)
+        return text
+
+    def common_path_suffix_unicode(self):
+        """CommonPathSuffixUnicode (variable):
+        An optional, NULL-terminated, Unicode string that is used to construct
+        the full path to the link item or link target by appending the string
+        in the CommonPathSuffixUnicode field. This field can be present only
+        if the VolumeIDAndLocalBasePath flag is set and the value of the
+        LinkInfoHeaderSize field is greater than or equal to 0x00000024.
+        """
+        if not self.common_path_suffix_offset_unicode():
+            return None
+
+        start = self.common_path_suffix_offset_unicode() + 4
+
+        binary = self._raw[start:]
+        # import pdb;pdb.set_trace()
+        text = self.text_processor.read_unicode_string(binary)
+        return text
