+++ tests/json/network_info.json
@@ -49,6 +49,7 @@
     },
     "link_info": {
         "common_network_relative_link_offset": 28,
+        "common_path_suffix": "A - LM METAL LIFT\\01.OBCHOD - BRO\u017dURY - Prodejn\u00ed a technick\u00e9 informace o produktech\\ETN\\ETN-Katalog-ENG\\Katalog ETN 10_2017\\Lift-programme\\ETN-lift programme 2017.pdf",
         "common_path_suffix_offset": 72,
         "link_info_flags": 2,
         "link_info_header_size": 28,
