+++ tests/json/readable_with_network_info.txt
@@ -77,6 +77,7 @@ Windows Shortcut Information:
       Local base path offset: 0
       Common network relative link offset: 28
       Common path suffix offset: 72
+      Common path suffix: A - LM METAL LIFT\01.OBCHOD - BROŽURY - Prodejní a technické informace o produktech\ETN\ETN-Katalog-ENG\Katalog ETN 10_2017\Lift-programme\ETN-lift programme 2017.pdf
          Common network relative link size: 44
          Common network relative link flags: 3
          Net name offset: 20
